Finalization of compliance profile management service
- cleanup of unused DAO classes
