+++ src/main/java/com/czertainly/core/dao/repository/ComplianceProfileRuleRepository.java
@@ -2,7 +2,6 @@ package com.czertainly.core.dao.repository;
 
 import com.czertainly.core.dao.entity.ComplianceProfile;
 import com.czertainly.core.dao.entity.ComplianceProfileRule;
-import com.czertainly.core.dao.entity.ComplianceRule;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -13,8 +12,6 @@ import java.util.UUID;
 public interface ComplianceProfileRuleRepository extends SecurityFilterRepository<ComplianceProfileRule, Long> {
 
     Optional<ComplianceProfileRule> findByUuid(UUID uuid);
-
-    Optional<ComplianceProfileRule> findByComplianceProfileAndComplianceRule(ComplianceProfile complianceProfile, ComplianceRule complianceRule);
 
     List<ComplianceProfileRule> findByUuidIn(List<UUID> uuid);
