+++ src/main/java/com/czertainly/core/service/ComplianceProfileService.java
@@ -1,22 +1,16 @@
 package com.czertainly.core.service;
 
 import com.czertainly.api.exception.*;
-import com.czertainly.api.model.client.compliance.v2.ComplianceProfileGroupsPatchRequestDto;
-import com.czertainly.api.model.client.compliance.v2.ComplianceProfileRequestDto;
-import com.czertainly.api.model.client.compliance.v2.ComplianceProfileRulesPatchRequestDto;
-import com.czertainly.api.model.client.compliance.v2.ComplianceProfileUpdateRequestDto;
+import com.czertainly.api.model.client.compliance.*;
+import com.czertainly.api.model.client.raprofile.SimplifiedRaProfileDto;
 import com.czertainly.api.model.common.BulkActionMessageDto;
-import com.czertainly.api.model.core.auth.Resource;
-import com.czertainly.api.model.core.compliance.v2.ComplianceGroupListDto;
-import com.czertainly.api.model.core.compliance.v2.ComplianceProfileDto;
-import com.czertainly.api.model.core.compliance.v2.ComplianceProfileListDto;
-import com.czertainly.api.model.core.compliance.v2.ComplianceRuleListDto;
-import com.czertainly.api.model.core.other.ResourceObjectDto;
+import com.czertainly.api.model.core.certificate.CertificateType;
+import com.czertainly.api.model.core.compliance.ComplianceProfileDto;
+import com.czertainly.api.model.core.compliance.ComplianceProfilesListDto;
 import com.czertainly.core.security.authz.SecuredUUID;
 import com.czertainly.core.security.authz.SecurityFilter;
 
 import java.util.List;
-import java.util.UUID;
 
 public interface ComplianceProfileService extends ResourceExtensionService {
     /**
@@ -24,7 +18,7 @@ public interface ComplianceProfileService extends ResourceExtensionService {
      *
      * @return List of compliance profiles
      */
-    List<ComplianceProfileListDto> listComplianceProfiles(SecurityFilter filter);
+    List<ComplianceProfilesListDto> listComplianceProfiles(SecurityFilter filter);
 
     /**
      * Get the details of a compliance profile
@@ -41,27 +35,70 @@ public interface ComplianceProfileService extends ResourceExtensionService {
      * @param request Request containing the attributes to create a new compliance profile. See {@link ComplianceProfileRequestDto}
      * @return DTO of the new compliance profile that was created
      * @throws AlreadyExistException Thrown when an existing compliance profile is found with the same name
+     * @throws NotFoundException     Thrown when a Rule or Group is not found
      * @throws ValidationException   Thrown when the attributes validations are failed for a rule in the request
      */
-    ComplianceProfileDto createComplianceProfile(ComplianceProfileRequestDto request) throws AlreadyExistException, ConnectorException, NotFoundException, AttributeException;
+    ComplianceProfileDto createComplianceProfile(ComplianceProfileRequestDto request) throws AlreadyExistException, NotFoundException, ValidationException, AttributeException, ConnectorException;
 
     /**
-     * Update compliance profile
+     * Add a rule to a compliance profile
      *
-     * @param request Request containing the attributes to update compliance profile. See {@link ComplianceProfileUpdateRequestDto}
-     * @return DTO of compliance profile that was updated
-     * @throws NotFoundException Thrown when the system cannot find the compliance profile for the given Uuid
-     * @throws ValidationException   Thrown when the attributes validations are failed for a rule in the request
+     * @param uuid    Uuid of the compliance provider
+     * @param request Parameters for adding a new rule to the compliance profile. See {@link ComplianceRuleAdditionRequestDto}
+     * @return Compliance Profile Dto
+     * @throws AlreadyExistException Thrown when the rule is already tagged with the Compliance Profile
+     * @throws NotFoundException     Thrown when unable to find the rule with the provided details
+     * @throws ValidationException   Thrown when the attribute validation fails for the given rule
      */
-    ComplianceProfileDto updateComplianceProfile(SecuredUUID uuid, ComplianceProfileUpdateRequestDto request) throws NotFoundException, ConnectorException, AttributeException;
+    ComplianceProfileRuleDto addRule(SecuredUUID uuid, ComplianceRuleAdditionRequestDto request) throws AlreadyExistException, NotFoundException, ValidationException, ConnectorException;
+
+    /**
+     * Remove a rule from a compliance profile
+     *
+     * @param uuid    Uuid of the compliance provider
+     * @param request Parameters required to remove a specific rule from the compliance profile
+     * @return Compliance Profile DTO
+     * @throws NotFoundException Thrown when the rule is not found with the profile
+     */
+    ComplianceProfileRuleDto removeRule(SecuredUUID uuid, ComplianceRuleDeletionRequestDto request) throws NotFoundException, ConnectorException;
+
+    /**
+     * Add a group to a compliance profile
+     *
+     * @param uuid    Uuid of the compliance provider
+     * @param request Parameters for adding a new group to the compliance profile. See {@link ComplianceGroupRequestDto}
+     * @return
+     * @throws AlreadyExistException Thrown when the selected group is already associated
+     */
+    ComplianceProfileDto addGroup(SecuredUUID uuid, ComplianceGroupRequestDto request) throws AlreadyExistException, NotFoundException, ConnectorException;
+
+    /**
+     * Delete a group from a compliance profile
+     *
+     * @param uuid    Uuid of the compliance provider
+     * @param request Parameters for deleting group to the compliance profile. See {@link ComplianceGroupRequestDto}
+     * @return Compliance Profile DTO
+     * @throws NotFoundException Thrown when the selected group is not found associated with the compliance profile
+     */
+    ComplianceProfileDto removeGroup(SecuredUUID uuid, ComplianceGroupRequestDto request) throws NotFoundException, ConnectorException;
+
+    /**
+     * Get the list of associated RA Profile to the compliance profile
+     *
+     * @param uuid Uuid of the compliance profile
+     * @return List of RA Profiles associated with the compliance profile. {@link SimplifiedRaProfileDto}
+     * @throws NotFoundException * @throws NotFoundException Thrown when a Rule or Group is not found
+     */
+    List<SimplifiedRaProfileDto> getAssociatedRAProfiles(SecuredUUID uuid);
 
     /**
      * Delete a compliance profile
      *
      * @param uuid UUID of the compliance profile
      * @throws NotFoundException   Thrown when the system is not able to find the compliance profile for the given UUID
+     * @throws ValidationException Thrown when there are any RA Profile association for the selected compliance profile
      */
-    void deleteComplianceProfile(SecuredUUID uuid) throws NotFoundException;
+    void deleteComplianceProfile(SecuredUUID uuid) throws NotFoundException, ValidationException;
 
     /**
      * Remove multiple compliance profiles
@@ -71,7 +108,7 @@ public interface ComplianceProfileService extends ResourceExtensionService {
      * @throws ValidationException Thrown when the profiles are dependencies for other objects
      * @throws NotFoundException   Thrown when a Rule or Group is not found
      */
-    List<BulkActionMessageDto> bulkDeleteComplianceProfiles(List<SecuredUUID> uuids);
+    List<BulkActionMessageDto> bulkDeleteComplianceProfiles(List<SecuredUUID> uuids) throws NotFoundException, ValidationException;
 
     /**
      * Remove compliance profiles forcefully. This methods makes removes the object dependencies and set them null.
@@ -82,100 +119,47 @@ public interface ComplianceProfileService extends ResourceExtensionService {
     List<BulkActionMessageDto> forceDeleteComplianceProfiles(List<SecuredUUID> uuids);
 
     /**
-     * List compliance rules by specified criteria
+     * List of all compliance rules for User Interface
      *
-     * @param connectorUuid connector UUID. When not provided, internal (workflow rules are listed)
-     * @param kind
-     * @param resource resource of the rule
-     * @param type
-     * @param format
-     * @return
-     * @throws NotFoundException
+     * @param complianceProviderUuid - UUID of the compliance provider
+     * @param kind                   Kind of the compliance provider
+     * @param certificateType        Type of the certificate for which the rules has to be fetched
+     * @return List of the rules for given connector and its kind
+     * @throws ConnectorException Thrown when there are issues related to connector communication
      */
-    List<ComplianceRuleListDto> getComplianceRules(UUID connectorUuid, String kind, Resource resource, String type, String format) throws NotFoundException, ConnectorException;
+    List<ComplianceRulesListResponseDto> getComplianceRules(String complianceProviderUuid, String kind, List<CertificateType> certificateType) throws NotFoundException, ConnectorException;
 
     /**
-     * List compliance groups by specified criteria
+     * List of all compliance groups from the compliance providers
      *
-     * @param connectorUuid
-     * @param kind
-     * @param resource resource of the group
-     * @return
-     * @throws NotFoundException
+     * @param complianceProviderUuid Uuid of the compliance provider
+     * @param kind                   Kind of the compliance provider
+     * @return List of compliance groups
+     * @throws ConnectorException Thrown when there are issues with the connector communication and operations
      */
-    List<ComplianceGroupListDto> getComplianceGroups(UUID connectorUuid, String kind, Resource resource) throws NotFoundException, ConnectorException;
+    List<ComplianceGroupsListResponseDto> getComplianceGroups(String complianceProviderUuid, String kind) throws NotFoundException, ConnectorException;
 
     /**
-     * @param groupUuid
-     * @param connectorUuid
-     * @param kind
-     * @return
-     * @throws NotFoundException
-     * @throws ConnectorException
-     */
-    List<ComplianceRuleListDto> getComplianceGroupRules(UUID groupUuid, UUID connectorUuid, String kind) throws NotFoundException, ConnectorException;
-
-    /**
-     * @param uuid
-     * @param request
-     */
-    void patchComplianceProfileRule(SecuredUUID uuid, ComplianceProfileRulesPatchRequestDto request) throws NotFoundException, ConnectorException;
-
-    /**
-     * @param uuid
-     * @param request
-     */
-    void patchComplianceProfileGroup(SecuredUUID uuid, ComplianceProfileGroupsPatchRequestDto request) throws ConnectorException, NotFoundException;
-
-    /**
-     * Get the list of associated resource objects to the compliance profile
+     * Associate a compliance profile to an RA Profile
      *
-     * @param uuid Uuid of the compliance profile
-     * @return List of resource objects associated with the compliance profile. {@link ResourceObjectDto}
-     * @throws NotFoundException * @throws NotFoundException Thrown when compliance profile is not found
+     * @param uuid       Uuid of the compliance profile
+     * @param raProfiles Uuid of the RA Profile. See {{@link RaProfileAssociationRequestDto}}
+     * @throws NotFoundException Thrown when either of the profiles are not found
      */
-    List<ResourceObjectDto> getAssociations(SecuredUUID uuid) throws NotFoundException;
+    void associateProfile(SecuredUUID uuid, RaProfileAssociationRequestDto raProfiles) throws NotFoundException;
 
     /**
-     * List compliance profiles associated with resource object
-     *
-     * @param resource
-     * @param associationObjectUuid
-     * @return
-     * @throws NotFoundException
-     */
-    List<ComplianceProfileListDto> getAssociatedComplianceProfiles(Resource resource, UUID associationObjectUuid);
-
-    /**
-     * Associate a compliance profile to resource object
-     *
-     * @param uuid uuid of compliance profile
-     * @param resource
-     * @param associationObjectUuid
-     * @throws NotFoundException
-     */
-    void associateComplianceProfile(SecuredUUID uuid, Resource resource, UUID associationObjectUuid) throws NotFoundException;
-
-    /**
-     * Disassociate a compliance profile from resource object
-     *
-     * @param uuid uuid of compliance profile
-     * @param resource
-     * @param associationObjectUuid
-     * @throws NotFoundException
-     */
-    void disassociateComplianceProfile(SecuredUUID uuid, Resource resource, UUID associationObjectUuid) throws NotFoundException;
-
-    /**
-     * Check the compliance for all objects associated with the compliance profiles
+     * Check the compliance for all the certificates associated with the compliance profiles
      *
      * @param uuids List of UUIDs of the compliance profiles
      */
     void checkCompliance(List<SecuredUUID> uuids);
 
     /**
-     * @param resource
-     * @param objectUuid
+     * Disassociate Compliance Profiles from RA Profiles
+     *
+     * @param uuid       Compliance Profile UUID
+     * @param raProfiles List of RA Profile UUIDs
      */
-    void checkResourceObjectCompliance(Resource resource, UUID objectUuid);
+    void disassociateProfile(SecuredUUID uuid, RaProfileAssociationRequestDto raProfiles) throws NotFoundException;
 }
